Gensym

Description

The code in this module is considered "unsafe". It should work ok in most circumstances, but there might be corner cases where it will cause trouble. Do reach out if you experience difficulties using this module.

Documentation

gensym :: String -> String

Generate a fresh variable based on the given one. The resulting variable has the form "xn" is a "uniqueness suffix".

Warning: This function relies on unsafePerformIO to avoid putting otherwise pure code into the IO monad.

Examples:

>>> gensym "X"
"X#2"

>>> gensym "X"
"X#3"

>>> gensym "Z"
"Z#4"