Types

Description

Types used in protoScheme.

This module defines the type language of protoScheme. Types have the following syntax:

<BaseType> ::= Integer
             | Real
             | Boolean

<Type> ::= <BaseType>
         | (-> <Type>+)
         | (Pair-of <Type> <Type>)
         | (List-of <Type>)
         | <TypeVariable>

Documentation

data BaseType

Base types

Constructors

TyInteger

TyReal

TyBoolean

data Type

protoScheme types

Constructors

TyBase BaseType
Base types

TyArrow [Type]
Function type

TyPair Type Type
Pair type

TyList Type
List type

TyVar TyVariable
Type variable

data Scheme

Type schemes for polymorphic types.

Constructors

All [TyVariable] Type

type TyVariable = String

Type variables

type Constraint = (Type, Type)

Constraints between types. The pair (ty1, ty2) represents a constraint ty1 = ty2.

type Substitution = [(TyVariable, Type)]

Type variable substitutions: sets of type-variable and type pairings.

fromSExpression :: Expr -> Result Type

Convert an s-expression into a type, if possible.

toSExpression :: Type -> Expr

Convert a type into an s-expression.

schemeFromSExpression :: Expr -> Result Scheme

Parse a scheme from an s-expression.

schemeToSExpression :: Scheme -> Expr

Convert a type scheme to an s-expression.

showType :: Type -> String

Pretty print a type as its s-expression representation

showScheme :: Scheme -> String

Pretty print a type as its s-expression representation

unify :: [Constraint] -> Result Substitution

Unify the given list of constraints

specializes :: Type -> Type -> Bool

Does the LHS type specialize the RHS type?

tySubst :: TyVariable -> Type -> Type -> Type

Substitute the given type for the given variable.

applySubst :: Substitution -> Type -> Type

Apply all substitutions in the given list to the given type

freeTyVars :: Type -> [TyVariable]

Free type variables in a type.

freeSchemeVars :: Scheme -> [TyVariable]

Free type variables in a type scheme.

generalizeType :: [TyVariable] -> Type -> Scheme

Generalize the given type, omitting the given variables.

instantiateScheme :: Scheme -> Type

Instantiate a type scheme with fresh type variables.

instantiateSchemeWith :: [TyVariable] -> Scheme -> Type

Instantiate a type scheme with the given type variables.